@@ -1,15 +1,16 @@
 plugins {
     id 'com.android.application'
+    id 'com.google.gms.google-services'
 }
 
 android {
     namespace 'com.example.journal'
-    compileSdk 33
+    compileSdk 34
 
     defaultConfig {
         applicationId "com.example.journal"
-        minSdk 33
-        targetSdk 33
+        minSdk 24
+        targetSdk 34
         versionCode 1
         versionName "1.0"
 
@@ -33,8 +34,19 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.6.1'
     implementation 'com.google.android.material:material:1.9.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
+    implementation platform('com.google.firebase:firebase-bom:32.7.2')
+    implementation 'com.google.firebase:firebase-auth'
+    implementation 'com.google.firebase:firebase-database'
+    implementation 'androidx.credentials:credentials:1.2.0'
+    implementation 'androidx.credentials:credentials-play-services-auth:1.2.0'
+    implementation 'com.google.android.libraries.identity.googleid:googleid:1.1.1'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
     implementation 'androidx.recyclerview:recyclerview:1.3.1'
+    implementation 'com.google.firebase:firebase-firestore-ktx'
+    implementation 'com.jakewharton.threetenabp:threetenabp:1.4.6'
+    implementation 'com.squareup.okhttp3:okhttp:4.12.0'
+    implementation 'org.json:json:20210307'
+
 }
